# docs/swagger.yaml
basePath: /
definitions:
  controller.AccountInput:
    properties:
      document_number:
        type: string
    type: object
  controller.TransactionInput:
    properties:
      account_id:
        type: integer
      amount:
        type: number
      operation_type_id:
        type: integer
    type: object
  models.Account:
    properties:
      account_id:
        type: integer
      document_number:
        type: string
    type: object
  models.Transaction:
    properties:
      account_id:
        type: integer
      amount:
        type: number
      event_date:
        type: string
      operation_type_id:
        type: integer
      transaction_id:
        type: integer
    type: object
host: localhost:8082
info:
  contact: {}
  description: This is a sample server for a bank API.
  title: Bank API
  version: "1.0"
paths:
  /accounts:
    post:
      consumes:
      - application/json
      description: Create a new account with the given document number
      parameters:
      - description: Account document number
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/controller.AccountInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new account
      tags:
      - accounts
  /accounts/{id}:
    get:
      consumes:
      - application/json
      description: Get information of an account by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Account not found
          schema:
            type: string
      summary: Get account information
      tags:
      - accounts
  /transactions:
    post:
      consumes:
      - application/json
      description: Create a new transaction with the given details
      parameters:
      - description: Transaction details
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/controller.TransactionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new transaction
      tags:
      - transactions
swagger: "2.0"